requirements.txt
paddleocr[doc-parser]
opencv-python-headless
Pillow
fastapi
uvicorn[standard]
python-multipart
